#745 - move link item interface to types, update imports
