+++ src/frontend/layouts/sidebar/nav-page-links/nav-page-links.tsx
@@ -5,14 +5,8 @@
 
 import { NavLink } from 'react-router-dom';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { IconProp } from '@fortawesome/fontawesome-svg-core';
+import { LinkItem } from '../../../../shared/types';
 import styles from './nav-page-links.module.css';
-
-export interface LinkItem {
-  name: string;
-  icon?: IconProp;
-  route: string;
-}
 
 interface NavPageLinkProps {
   linkItems: LinkItem[];
